Find references to procs as [sub calls]
Recognise the word lists and traverse into them when finding the cursor
position. This recognises the position in things like:

    Outer arg [Inner arg arg]
                ^ here

As the first word of the Inner call, as opposed to the 3rd word of the
Outer call. As with the Parser and the Indexer we split the traversal
into TraverseWord and TraverseScript and handle appropriately.

We should really generalise that, with some sort of visitor.