Sort of support multiple definitions of procs
We use a dodgy algorith that's very untested to determine the best fit
proc definition for a given Call site. The idea is to try and match the
arguments specified to the arguments required.

It might/would be better to just use the latest one defined at the
source location but that's actually kinda fiddly.