Dodgy background thread for indexing
We start up 4 threads in a thread pool (using io_context) and create a
strand for executing parsing operations sequentially in that pool.

This is all data race central, because there's absolutely no
synchronisation on the index itself, and the stl containers within are
not threadsafe, but for now it shows how it's possible to schedule stuff
on a thread pool and to co_await the result of such operations, which is
what i was trying to acheive experimentally.

Also stop using a global static for the server. It's time.